third_party/blink/web_tests/external/wpt/scroll-to-text-fragment/non-html-documents.html

<!doctype html>
<title>Allow text fragments in HTML documents only</title>
<meta charset=utf-8>
<link rel="help" href="https://wicg.github.io/ScrollToTextFragment/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/common/utils.js"></script>
<script src="stash.js"></script>

<script>

function loadEvent(win) {
  return new Promise((resolve) => {
    if (win.document.readyState == 'complete')
      resolve();
    else
      win.onload = resolve;
  });
}

function rAF(win) {
  return new Promise((resolve) => {
    win.requestAnimationFrame(resolve);
  });
}

const test_cases = [
  'non-html.css',
  'non-html.js',
  'non-html.json',
  'non-html.txt',
  'non-html.xml'
];

for (let test_case of test_cases) {
  promise_test(async function (t) {
    let popup = null;
    test_driver.bless('Open a URL with a text fragment directive', () => {
      popup = window.open(`resources/${test_case}#:~:text=target`, '_blank', 'width=300,height=300');
    });

    await t.step_wait(() => popup != null, "Opened popup window");
    await loadEvent(popup);

    // rAF twice in case there is any asynchronicity in scrolling to the target.
    await rAF(popup);
    await rAF(popup);

    assert_equals(popup.scrollY, 0);
    popup.close();
  }, `Text directive blocked in ${test_case}`);
}

</script>
